+++ components/manualBox/manualBox.types.ts
