+++ components/manualBox/myManualLayout.module.scss
